app/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.1.8
coverage==7.10.7
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.2.1
tomli==2.3.0
typing_extensions==4.15.0
Werkzeug==3.1.3
zipp==3.23.0
